Fix reused object in variable parsing.
The default value object was created once on compilation and reused on every query.
This was a mistake that is harmless when all values are always specified.

When one of the inputs is sometimes not present, then this bugs show up with the previous value that it was set to.

This PR correctly creates the object inside the instantiated function.