Reuse arrays
The intermediate array allocated to generate the final shape is pointless if the users do not plan on using the previous one.

This allows saving pointless array allocations.
It is disabled by default since it is an dangerous option for the cases where the user does require the original array to be kept.