Fix a typechecking bug that newer TypeScripts catch
Newer typescripts now catch a valid issue where a property was expected to be defined while being undefined. After bumping the compiler, this came up:

```
$ yarn run tsc --noEmit
src/ast.ts:374:52 - error TS2322: Type 'undefined' is not assignable to type 'ObjectPath'.

374     return { value: null, variables: [{ valueNode, path: undefined }] };
                                                       ~~~~

  src/ast.ts:338:3
    338   path: ObjectPath;
          ~~~~
    The expected type comes from property 'path' which is declared here on type 'MissingVariablePath'
```

Since this code has been like this for a while and works just fine, this just updates the types to reflect that the `path` of a `MissingVariablePath` can be undefined, and updates the functions that use it to be a bit simpler when it is undefined.