Move object initalisation behind a feature toggle
Moved the initialisation of __internalShouldIncludePath under field node
to happen behind the configuration option
compilationContext.options.useExperimentalPathBasedSkipInclude.

The context for the change is that updating from 0.8.0 to 0.8.2 has
shown increased Garbage Collector rates in use cases with big queries,
and this change could be used to verify the hypothesis whether the
initialisation of __internalShouldIncludePath is a cause for the GC
runs.